backend/pyproject.toml
[project]
name = "bouldering-app-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.0.0",
    "supabase>=2.0.0",
    "pyjwt>=2.8.0",
    "openai>=1.0.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "openpyxl>=3.1.0",
    "google-genai>=1.0.0",
    "yt-dlp>=2024.0.0",
    "youtube-transcript-api>=0.6.2",
]

[tool.uv]
dev-dependencies = [
    "pre-commit>=4.0.0",
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.9.0",
    "ty>=0.0.1a0",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "UP",
    "B",
]

[tool.ruff.lint.per-file-ignores]
"src/app/dependencies/*.py" = ["B008"]
"src/app/routers/*.py" = ["B008"]

[tool.ruff.format]
quote-style = "double"

[tool.ty.environment]
python-version = "3.11"
extra-paths = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
markers = [
    "integration: DB integration tests (require supabase start)",
]
